JAXB entity serialization; Import GML service operation.

+++ eenvplus-sdi-model/src/main/java/be/vmm/eenvplus/sdi/model/type/ReferenceSerializer.java
@@ -18,7 +18,7 @@ public class ReferenceSerializer extends JsonSerializer<Reference> {
 			JsonProcessingException {
 
 		if (reference.getType() == ReferenceType.id)
-			generator.writeNumber(reference.getValue());
+			generator.writeNumber((Long) reference.getValue());
 		else
 			generator.writeString(reference.toString());
